feat(auth): protect post & comment routes with ownership checks

+++ controllers/comment.js
@@ -1,6 +1,7 @@
 import * as commentModel from "../models/comment.js";
 import passport from "passport";
 import { body, matchedData, validationResult } from "express-validator";
+import { CustomAuthorizationError } from "../errors/CustomAuthorizationError.js";
 
 const validations = [
   body("text")
@@ -30,6 +31,7 @@ const createComment = [
 ];
 
 const updateComment = [
+  passport.authenticate("jwt", { session: false }),
   validations,
   async function (req, res) {
     const commentId = req.params.commentId;
@@ -39,16 +41,34 @@ const updateComment = [
       res.status(400).json({ error: errors.array() });
     }
     const { text } = matchedData(req);
+    const commentTobeUpdated = await commentModel.getComment(commentId);
 
-    const comment = await commentModel.updateComment({ id: commentId, text });
-    res.json({ comment });
+    if (req.user.id == commentTobeUpdated.authorId) {
+      console.log(commentTobeUpdated);
+      const comment = await commentModel.updateComment({ id: commentId, text });
+
+      return res.json({ comment });
+    }
+    else{
+        throw new CustomAuthorizationError("Access denied")
+    }
   },
 ];
 
-async function deleteComment(req, res) {
-  await commentModel.deleteComment(req.params.commentId);
-  res.json({ message: "Comment deleted successfully" });
-}
+const deleteComment = [
+  passport.authenticate("jwt", { session: false }),
+  async function (req, res) {
+    const commentId = req.params.commentId;
+    const commentTobeUpdated = await commentModel.getComment(commentId);
+
+    if (req.user.id == commentTobeUpdated.authorId) {
+      await commentModel.deleteComment(req.params.commentId);
+      res.json({ message: "Comment deleted successfully" });
+    } else {
+      throw new CustomAuthorizationError("Access denied")
+    }
+  },
+];
 
 async function likeComment(req, res) {
   const comment = await commentModel.likeComment(req.params.commentId);
